djangobackend/requirements.txt
asgiref==3.8.1
coverage==7.6.7
Django==5.1.3
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
pillow==11.0.0
psycopg2-binary==2.9.10
PyJWT==2.10.0
python-decouple==3.8
sqlparse==0.5.2
typing_extensions==4.12.2
